# [2] graph/db: batch-fetch node data