This is the second PR in the graph-SQL performance improvement series. This one is aimed at improving the performance of the `ForEachNode` method which iterates through all the nodes in the graph DB. 
This is one by:
1) using batched pre-loading to fetch the addresses/features/extra-types for a set of nodes rather than fetching one by one. 
2) manipulating default paged query size. NOTE: an action item from this is to have different defaults for SQLite & Postgres since postgres seems to work well with larger page sizes but not so for sqlite. 

part of part of https://github.com/lightningnetwork/lnd/issues/9795

# Benchmark Results


native-sqlite: Batching  initially made it worse (-114.8%), but page size changes improved it from batching (+63.8%), overall +22.1% from baseline.
native-postgres: Batching improved it dramatically (+92.3%), page size changes made it slightly worse (-51.3%), but overall still +88.3% from baseline


## ForEachNode Performance

| Database | Baseline (ns/op) | Batching (ns/op) | Batching % | Page Size (ns/op) | Page Size % | Overall % |
|----------|------------------|------------------|------------|-------------------|-------------|-----------|
| kvdb-sqlite | 207,190,792 | 207,136,758 | N/A | 211,486,375 | N/A | N/A |
| native-sqlite | 359,364,150 | 772,132,008 | -114.8% | 279,867,708 | +63.8% | +22.1% |
| kvdb-postgres | 625,893,875 | 662,636,250 | N/A | 630,299,683 | N/A | N/A |
| native-postgres | 1,244,927,533 | 96,298,042 | +92.3% | 145,737,983 | -51.3% | +88.3% |

## NodeUpdatesInHorizon Performance

| Database | Baseline (ns/op) | Batching (ns/op) | Batching % | Page Size (ns/op) | Page Size % | Overall % |
|----------|------------------|------------------|------------|-------------------|-------------|-----------|
| kvdb-sqlite | 297,871,125 | 300,307,533 | N/A | 298,996,058 | N/A | N/A |
| native-sqlite | 478,691,733 | 578,211,708 | -20.8% | 238,166,717 | +58.8% | +50.2% |
| kvdb-postgres | 893,502,800 | 1,028,758,267 | N/A | 893,580,392 | N/A | N/A |
| native-postgres | 1,275,379,175 | 79,515,367 | +93.8% | 217,169,542 | -173.1% | +83.0% |